+++ examples/agentic-editing/CHANGELOG.md
@@ -1,5 +1,19 @@
 # agentic-editing-example
 
+## 0.0.1-alpha.9
+
+### Patch Changes
+
+- Updated dependencies [[`de6ad86`](https://github.com/mattppal/mcut/commit/de6ad86850e763a48f50e4800329e66ba5551928)]:
+  - @mcut/timeline@0.1.0-alpha.10
+
+## 0.0.1-alpha.8
+
+### Patch Changes
+
+- Updated dependencies [[`9f1ebce`](https://github.com/mattppal/mcut/commit/9f1ebce07c0d449e6818951fabb61668b27852e2)]:
+  - @mcut/timeline@0.1.0-alpha.9
+
 ## 0.0.1-alpha.7
 
 ### Patch Changes