+++ examples/mcp-server/CHANGELOG.md
@@ -1,5 +1,21 @@
 # mcp-server-example
 
+## 0.0.1-alpha.15
+
+### Patch Changes
+
+- Updated dependencies [[`de6ad86`](https://github.com/mattppal/mcut/commit/de6ad86850e763a48f50e4800329e66ba5551928), [`de6ad86`](https://github.com/mattppal/mcut/commit/de6ad86850e763a48f50e4800329e66ba5551928)]:
+  - @mcut/mcp-server@0.1.0-alpha.16
+  - @mcut/timeline@0.1.0-alpha.10
+
+## 0.0.1-alpha.14
+
+### Patch Changes
+
+- Updated dependencies [[`888fa37`](https://github.com/mattppal/mcut/commit/888fa379f17f4607a25ef8f73ff75f09e41d7c05), [`9f1ebce`](https://github.com/mattppal/mcut/commit/9f1ebce07c0d449e6818951fabb61668b27852e2), [`9f1ebce`](https://github.com/mattppal/mcut/commit/9f1ebce07c0d449e6818951fabb61668b27852e2)]:
+  - @mcut/mcp-server@0.1.0-alpha.15
+  - @mcut/timeline@0.1.0-alpha.9
+
 ## 0.0.1-alpha.13
 
 ### Patch Changes